requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
huggingface-hub[hf_xet]==0.34.4
requests>=2.31.0
pydantic>=2.5.0
torch==2.8.0
transformers==4.56.1
accelerate==1.10.1
llama-cpp-python[server]>=0.2.0
psutil>=5.9.0